added limit and scalability mechanism to thumbnailer
Modified thumbnailer to distribute cached files to number of
subfolders. There is now also a limit for maximum number
of cached items and maximum size of the cache (default values
are 100k items or total of 50Mb).

Subfolder name is constructed from cryptographic hash (which is
already calculated), there is practically no performance hit for
read operations. There is extra file access for cache writes,
because size and count is cached to separate text file. If that
book keeping file is missing or contains corrupted data, then it is
initialized to real situation based on actual files in the cache.

Signed-off-by: Antti Seppälä <antseppa@gmail.com>