Applying patch from Chris Adams <chris.adams@jollamobile.com>
Changing methods for phone number and email address that set their type to be index based to the list of numbers that the user has.

This has the downside that we emit a type changed signal for each type change, instead of the previous approach where we could assign the types in one call.