Implement sort by field for the proxy model.
List can be sorted either by first name or last name. This setting can be either set from SeasideProxyModel (via QML) or it can also be set and changed via GConf.

The sorting information is stored in SeasidePeopleModel which connects signals to SeasidePeople that will react to sorting being changed.

This commit will also deffer the creation of the display label until it is needed. This way we prevent it from being created twice, in case we initially want to have last name first.