Resolve display label in a more consistent manner
The qtcontacts-tracker engine uses the QContactGlobalPresence detail
to generate display label information (see
displaylabelgenerate.cpp:267) however that wasn't mirrored in the
SeasidePerson display label generation code.  This commit adds code
to grab the imNicknames from the imAddresses associated with the
contact, and saves them as QContactPresence details, which can be
used to generate display label information in a more conformant manner.